+++ pyTooling/GenericPath/URL.py
@@ -256,17 +256,6 @@ class Path(PathMixIn):
 	ROOT_DELIMITER:    ClassVar[str] = "/"                #: Delimiter symbol in URLs between root and first element.
 	ELEMENT_TYPE:      ClassVar[type[Element]] = Element  #: Type an element of a URL's path has.
 
-	@classmethod
-	def Parse(cls, path: str, root: Nullable[Host] = None) -> Path:
-		"""
-		Parse a string into a URL path.
-
-		:param path: The path portion of a URL.
-		:param root: Optional, host the path is relative to.
-		:returns:    The parsed path.
-		"""
-		return super().Parse(path, root, cls, Element)
-
 
 @export
 class URL: